requirement.txt
Flask==2.3.2
requests==2.31.0
wolframalpha==4.0.0
openai==0.27.0
nltk==3.8.1
python-dotenv==1.0.0